refact(ProductSelector): move into parsing standalone/additional quotes
